Clean up some warnings, and fix missing define in spiPreInit

+++ src/main/flight/servos.c
@@ -68,7 +68,7 @@ void pgResetFn_servoConfig(servoConfig_t *servoConfig)
     servoConfig->dev.ioTags[0] = IO_TAG(SERVO1_PIN);
 #else
     int servoIndex = 0;
-    for (int i = 0; i < USABLE_TIMER_CHANNEL_COUNT && servoIndex < MAX_SUPPORTED_SERVOS; i++) {
+    for (int i = 0; i < (int)USABLE_TIMER_CHANNEL_COUNT && servoIndex < MAX_SUPPORTED_SERVOS; i++) {
         if (timerHardware[i].usageFlags & TIM_USE_SERVO) {
             servoConfig->dev.ioTags[servoIndex] = timerHardware[i].tag;
             servoIndex++;
